Moves and renames some samples components more appropriately
